Fix config error. Move files around
