--- lua/plugins.lua
@@ -1,4 +1,0 @@
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
-vim.opt.rtp:prepend(lazypath)
-
-require('lazy').setup('plugins')
