+++ lua/plugins/init.lua
@@ -1,11 +1,4 @@
 return {
-    {
-        'nvim-treesitter/nvim-treesitter', 
-        build = ":TSUpdate",
-    },
-
-    {
-        'stevearc/dressing.nvim',
-        event = "VeryLazy",
-    },
+    { 'nvim-treesitter/nvim-treesitter', build = ":TSUpdate" },
+    { 'stevearc/dressing.nvim', event = "VeryLazy" },
 }
