+++ lua/plugins/treesitter.lua
@@ -1,7 +1,7 @@
 return {
     require'nvim-treesitter.configs'.setup {
         -- A list of parser names, or "all" (the five listed parsers should always be installed)
-        ensure_installed = { "c", "cpp", "lua", "rust", "vim", "vimdoc", "python" },
+        --ensure_installed = { "c", "cpp", "lua", "rust", "vim", "vimdoc", "python" },
 
         -- Install parsers synchronously (only applied to `ensure_installed`)
         sync_install = false,
